USB: MUSB: Do not disable session bit in musb_do_idle()
In OTG/HOST MUSB mode, When the Micro A connector used,
VBUS is turned on and session bit set. When the device
is connected, enumeration goes through. When the device
disconnected from the other end of the connector (ID is
still grounded), link will detect the disconnect and end
the session. When the device is connected back, there
are no events generated in the TWL6030-usb, and link is
already down. So the device is not detected.

Removed the session bit disable code which will recognize
the connect of the device.

Change-Id: I8b89048dc14065e91c7cf5830de8cbbd5a40f4f2
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>