usb: musb: Do not cut clock on suspend interrupt
The USB CV Mass Storage Test cases fail with an error "Device not
found" when the PHY clocks are cut in SUSPEND interrupt.

When the USB cable is connected, a wake lock is held; hence there
is no need to cut the clock in suspend interrupt.
This patch removes cutting and enabling of clock in SUSPEND and
RESUME interrupt respectively.

[Moiz] This patch also fixes the issue of gadget enumeration
in full-speed.

Change-Id: Id72ded67af1e16d12ecea6c2edc7cecac424ab08
Signed-off-by: Mohan V <mohanv@ti.com>
Acked-by: Moiz Sonasath <m-sonasath@ti.com>