OMAP4: DSI/V4L2 Fix for 60fps rendering on primary DS1
It was found that we were using TE enable instead of TE start to
send out the data for primary display. For manual update mode we should
be using TE start as recommended. In TE enable mode we will loose time
waiting for the TE interrupt to arrive from the panel which will occur
at 60 fps i.e 15 msec boundary. In case of TE start we can issue another
transfer after Framedone from the dispc. This also garantees no tear
since we wait for Framedone interrupt.

Change-Id: Ibbaefeea1c5474a90f9042d40280f6dc2d6ceb73
Signed-off-by: Dandawate Saket <dsaket@ti.com>