OMPA4: Set override mode for SRAM LDO'S bias adjusment.
For some ES2.2 silicons, efused values for BGAP and SDLDO
trim where not programmed.
This may cause system instabilities at low opp's.

As a workaround, if the silicon is not efused,
we can set overide mode for SLDO trim, and program
the maximum multiplication factor. This, will
ensure a proper voltage for the corresponding SLDO.

Change-Id: I3feb359599a6dd062be4ce6ec0fa8acb8ba22f9a
Signed-off-by: Axel Haslam <axelhaslam@ti.com>