USB: MUSB: Move the L3 constarint calls to Work Queue
Having the L3 constarint calls in the musb interrupt
handler was causing:
BUG: sleeping function called from invalid context at
kernel/mutex.c:94

This is because 'omap_pm_set_min_bus_tput' calls
'mutex_lock' which can sleep till it gets the lock.
But in interrupt context this is not allowed.

So, this patch moves the L3 contraint calls on attach/
detach into the musb work queue which is scheduled from
the musb interrupt handler.

Change-Id: I3da9a5bad32e4eb38a8ca5a1f4c6aee397bcdf8c
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>