drivers:staging:ti-st: fix FM get operations
FM character device driver converts the incoming CH8 data from the chip
into HCI-Event data as expected from the FM stack (MCPv2.5).
During this convertion, the parameter length field or rlen is not made
use of in HCI-Event packet and is always valued at 0.

This patch trims the size of the HCI-Event sent up to the FM stack.

Change-Id: I9729bf3f5204d9a094f7e3b9391b74ba6e702f31
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>