USB: Mass-storage: Hold 800 MHz MPU freq constriaint
This patch holds a 800 MHz MPU freq constarint
when a SD card is mounted on the host PC. This
is needed for better mass-storage USB-SD throughput
numbers.

It has been observed that without this patch,
the load on MPU for the mass-storage use-case
is not not sufficient and hence MPU stays
at lower frequencies, thereby causing a
throughput drop.

Change-Id: Ie0a252f66ea944ffa5a0510c78bd65c578d2619b
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>