OMAP4:Turn off VMMC LDO
Remove the always_on configuration of VMMC ldo so that it
can be turned off when not used.Since the ldo is turned off
a 1ms delay is required in the power up sequence so that
pbias can ramp up to the correct voltage.

In the suspend scenario ldo is forcibly turned off.In the
idle case the ldo is turned off only if there is no activity
on MMC/SD for 8Sec.

Change-Id: I33fa51b1cd47bfb5ec9ee241fbdbc442249c0f9b
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>