USB: TWL: Ignore ID-GND interrupts for Peripheral-only mode
During the USB Back Voltage testing in Peripheral-only mode,
on blaze boards it was observed that when you connect the usb
cable to the test fixture, it was generating spurious ID GND
interrupts and hence as the USB OTG SW stack works on cable-
based switching, it was causing the OTG controller to switch
its role of HOST and supply VBUS, causing the tests to fail.

With this patch we ignore any ID GND interrupts if we are
operating in PERIPHERAL-only mode.

Change-Id: I35454abcbbd3c890f69b5f4fa8a783b9b77ece9c
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>