OMAP: clock: add pre/post clock notifiers
This patch was adapted from Paul Walmsley's original patch introducting
clock notifiers: https://patchwork.kernel.org/patch/13320/

The original comment message is below:

"Call clock notifier blocks before and after rate and parent changes.
If the rate change fails for any reason after the notifier is called with
a pre-rate-change message, call the notifier again with an abort message.

Pre-rate-change notifiers are passed the current clock rate and the
desired clock rate.  For this to work with parent changes, additional
architecture-specific support was needed to predict the clock rate
after the parent change, so a "clk_round_rate_parent" function pointer
was added to struct clk_functions for this purpose, and this was
defined for the OMAP2/3 architecture.

(To minimize performance and code impact, post-rate-change notifiers,
however, are passed only the desired clock rate, i.e., the clock rate
after the rate change.)

The notifiers are called even if the clock rate does not change.  This
is because reprogramming a clock's parent or rate may briefly disrupt
the clock."

Changes to the original patchset include using a new "speculate" function
pointer to walk the list for the PRE change notifier instead of a temporary
storage in struct clk, as well as re-use of some existing clock framework
APIs for determining fieldvals properly.

Change-Id: I4299a6d22735f25083d66e1602c346a54b40a79d
Signed-off-by: Mike Turquette <mturquette@ti.com>