OMAP4: clock: don't touch CM_SCALE_FCLK in LP mode
In DPLL cascading mode we run all of the PER functional clocks at the rate
closest to their "normal" output.  In both the "normal" case and the DPLL
cascading scenario we want CM_SCALE_FCLK to be 0, which will run the PER
fclks at their full rate.  As such, no need to touch this register in DPLL
cascading code as 0 is the default.

Change-Id: I11982bce16e41753ea261fb9266edab682f93c3e
Signed-off-by: Mike Turquette <mturquette@ti.com>