OMAP4: PM: Make EMIF interrupts wakeup capable.
EMIF Temperature Alert Interrupt may be pending
when ever the system is going to suspend and a corner case
arises when the MPU has already gone to OFF Mode and the EMIF
is still ON as the Interrupt is pending. This
may cause the CORE to remain ON, and prevent
DEVICE_OFF.
Resolution is done by making the EMIF Interrupts wake up capable.

Change-Id: I9ff7d34c293eaface5c02007e50efe1ba2492573
Signed-off-by: Reddy, Jagadeeshwar <jagadeeshwar.reddy@ti.com>
Signed-off-by: Axel Haslam <axelhaslam@ti.com>