misc: twl6040-vib: Add errata for false overcurrent detection
It's possible to have false overcurrent detection when vibrator drivers
are enabled. TWL6040 reacts to the overcurrent detection event by
disabling vibrator outputs and generates a vibrator interrupt VIBINT.
To prevent false overcurrent detection, overcurrent protection is disabled
for at least 2.5ms to allow vibrator driver to stabilize. The period of
time overcurrent detection is disabled is very short and the probability
for an actual overcurrent is negligible.

Change-Id: I9ab7d05572b58c69e033edc6821277764b6250c6
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Axel Castaneda Gonzalez <x0055901@ti.com>