OMAP4: DSSCOMP: Fix to signal the dropped compositions
Description# The compositions which are marked BUSY in setupmgr method,
need to signalled once they are processed. Earlier the signal was done only
in the release callback, but this was missing the compositions which were
never applied. This patch moves the signal logic from release callback to
composition drop which takes care of released as well as dropped composition

Change-Id: I96fcabdfb1857dc5e9f8176a1053cdfaa4e8e25f
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>