OMAP4: DSSCOMP: Fix to not to alter composition in validation
Description# The return value from the validate method is captured in a
local variable instead of altering the correct composition.
Earlier the composition that was beign validated was modified to capture
errcode. This would make the composition NULL if the validate fails (means
compposition is not in expected state). This would result in kernel crashes
if the was accessed by the caller.

Change-Id: I70235a7b0db6ca0a0c4592de2dd372048c3b7b02
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>